% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{CalModMatrix}
\alias{CalModMatrix}
\title{Create a Calibration Model Matrix Object}
\usage{
CalModMatrix(calfit, ...)
}
\arguments{
\item{calfit}{a fitted linear model object from \link[stats]{lm} or a
named list containing 'coef': the coefficients of the calibration model
c(lambda0, lambdaxstar, lambdaz) and a non-mandatory 'vcov': the
variance--covariance matrix of the coefficients}

\item{substitute}{a vector containing the error-prone measure}
}
\value{
\code{CalModMatrix} returns an object of \link[base]{class}
"CalModMatrix".

An object of class \code{CalModMatrix} is a list containing the substitute
variable and the coefficients of the calibration model and the
variance--covariance matrix of the coefficients of calibration model and the
calibration model matrix. A \code{CalModMatrix} is typically used in
\link[mecor]{mecor} in case of measurement error in the independent variable.
The CalModMatrix is implemented in \code{mecor} to provide measurement error
correction using external data and for conducting sensitivity analyses. A
\code{CalModMatrix} object has attributes input (the name of the substitute
variable) and call (the matched call).
}
\description{
This function creates a calibration model matrix object, usually used as the
independent variable in a regression if one wants to correct for
the measurement error in that variable, by defining the coefficients of the
calibration matrix.
}
\examples{
## measurement error in dependent variable
data(ivs)
calmod_fit <- lm(X ~ X_star + Z, data = ivs)
calmod_coef <- coefficients(calmod_fit)
calmod_vcov <- vcov(calmod_fit)
calmod_matrix <- with (ivs, CalModMatrix(substitute = X_star,
                                         calfit = calmod_fit))
# identical to
calmod_matrix <- with (ivs, CalModMatrix(substitute = X_star,
                                         calfit = list(coef = calmod_coef,
                                                        vcov = calmod_vcov)))
}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
