% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccs}
\alias{ccs}
\title{Covariate-Calibration Study}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
  \item{Y}{outcome, continuous}
  \item{X_star}{error-prone exposure with systematic measurement error,
  continuous}
  \item{Z}{covariate, continuous}
  \item{X_star_1}{first replicate of error-prone exposure with classical
  measurement error, continuous}
  \item{X_star_2}{second replicate of error-prone exposure with classical
  measurement error, continuous}
}
}
\usage{
ccs
}
\description{
A simulated dataset containing 1000 observations of the outcome Y, the
error-prone exposure X_star with systematic measurement error, the covariate
Z and two replicates measures X1_star and X2_star of the exposure with
classical measurement error. The two replicates are observed in the first
500 study participants.
}
\examples{
data("ccs", package = "mecor")
}
\keyword{datasets}
