% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecvs}
\alias{ecvs}
\title{External Covariate-Validation Study}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{X}{gold standard exposure, continuous}
  \item{X_star}{error-prone exposure, continuous}
  \item{Z}{covariate, continuous}
}
}
\usage{
ecvs
}
\description{
A simulated dataset containing 100 observations of the gold standard
measurement of the exposure X, the error-prone exposure X_star and the
covariate Z. The outcome Y is not observed in the study population.
To be used in combination with the dataset \link[mecor]{icvs}.
}
\examples{
data("ecvs", package = "mecor")
}
\keyword{datasets}
