% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eovs}
\alias{eovs}
\title{External Outcome-Validation Study}
\format{
A data frame with 100 rows and 2 variables:
\describe{
  \item{Y}{gold standard outcome, continuous}
  \item{Y_star}{error-prone outcome, continuous}
}
}
\usage{
eovs
}
\description{
A simulated dataset containing 100 observations of the gold standard
measurement of outcome Y and the error-prone outcome Y_star. The covariates X
and Z are not observed in the study population. To be used in combination
with the dataset \link[mecor]{iovs}.
}
\examples{
data("eovs", package = "mecor")
}
\keyword{datasets}
