% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icvs}
\alias{icvs}
\title{Internal Covariate-Validation Study}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{Y}{outcome, continuous}
  \item{X_star}{error-prone exposure, continuous}
  \item{Z}{covariate, continuous}
  \item{X}{gold standard exposure, continuous}
}
}
\usage{
icvs
}
\description{
A simulated dataset containing 1000 observations of the outcome Y, the
error-prone exposure X_star and the covariate Z. The gold standard exposure
X is observed in approximately 25\% of the study population
}
\examples{
data("icvs", package = "mecor")
}
\keyword{datasets}
