% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iovs}
\alias{iovs}
\title{Internal Outcome-Validation Study}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{Y_star}{error-prone outcome, continuous}
  \item{X}{exposure, continuous}
  \item{Z}{covariate, continuous}
  \item{Y}{gold standard outcome, continuous}
}
#' @examples
data("iovs", package = "mecor")
}
\usage{
iovs
}
\description{
A simulated dataset containing 1000 observations of the error-prone outcome
Y_star, the exposure X and the covariate Z. The gold standard outcome Y is
observed in approximately 25\% of the study population
}
\keyword{datasets}
