% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iovs_diff}
\alias{iovs_diff}
\title{Internal Outcome Validation Study}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
  \item{Y_star}{error-prone outcome, continuous}
  \item{X}{exposure, binary}
  \item{Y}{gold standard outcome, continuous}
}
}
\usage{
iovs_diff
}
\description{
A simulated dataset containing 1000 observations of the error-prone outcome
Y_star with differential measurement error, and the binary exposure X. The
gold standard  outcome Y is observed in approximately 25\% of the study
population
}
\examples{
data("iovs_diff", package = "mecor")
}
\keyword{datasets}
