% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_me.R
\name{lm_me}
\alias{lm_me}
\title{Run linear regression with correction method}
\usage{
lm_me(x, y, w, dataset, calset, form_of_me = cme,
  correction_method = zero_variance)
}
\arguments{
\item{x}{A vector with the exposure variable}

\item{y}{A vector with the outcome variable}

\item{w}{A vector with the surrogate outcome variable}

\item{dataset}{A dataframe with the dataset}

\item{calset}{A dataframe with the calibration dataset}

\item{form_of_me}{Form of measurement error (default is classical)}

\item{correction_method}{Correction method (default is zero_variance)}
}
\value{
This function returns (..).
}
\description{
This is a function that runs a linear regression with a correction method.
You can fill your function using the \code{x}.
}
\examples{
lm_me(x, y, w)
lm_me(x, y, w, form_of_me = dme)

}
