% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasErrorExt.R
\name{MeasErrorExt}
\alias{MeasErrorExt}
\title{Create External Measurement Error Object}
\usage{
MeasErrorExt(substitute, model)
}
\arguments{
\item{substitute}{a vector containing the error-prone measure}

\item{model}{a fitted linear model of class \link[stats]{lm} or a list
containing the coefficients of the calibration model or measurement error
model}
}
\value{
\code{MeasErrorExt} returns an object of \link[base]{class}
"MeasErrorExt".

An object of class \code{MeasErrorExt} is a list containing the substitute
variable and the fitted calibration model or measurement error model and has
attributes input (the name of the substitute variable) and call (the matched
call).
}
\description{
This function creates an external measurement error object, usually used as
the independent or dependent variable in a regression if one wants to correct
for the measurement error in that variable using external data or externally
estimated coefficients of the calibration model (independent variable) or
measurement error model (depedent variable)
}
\examples{
## measurement error in independent variable
data(ivs)
calmod_fit <- lm(X ~ X_star + Z, data = ivs)
calmod_coef <- coefficients(calmod_fit)
calmod_vcov <- vcov(calmod_fit)
me_ext <- with (ivs, MeasErrorExt(substitute = X_star,
                                  model = calmod_fit))
# identical to
me_ext <- with (ivs, MeasErrorExt(substitute = X_star,
                                  model = list(coef = calmod_coef,
                                               vcov = calmod_vcov)))
## measurement error in dependent variable
data(ivs_o)
measerror_fit <- lm(Y_star ~ Y, data = ivs_o)
measerror_coef <- coefficients(measerror_fit)
measerror_vcov <- vcov(measerror_fit)
me_ext <- with (ivs_o, MeasErrorExt(substitute = Y_star,
                                    model = measerror_fit))
# identical to:
me_ext <- with (ivs_o, MeasErrorExt(substitute = Y_star,
                                    model = list(coef = measerror_coef))
me_ext <- with (ivs_o, MeasErrorExt(substitute = Y_star,
                                    model = list(coef = c(0, 0.5))))
}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
