% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{MeasErrorMatrix}
\alias{MeasErrorMatrix}
\title{Create a Measurement Error Matrix Object}
\usage{
MeasErrorMatrix(mefit, ...)
}
\arguments{
\item{mefit}{a fitted linear model object from \link[stats]{lm} or a named list
containing 'coef': the coefficients of the measurement error
model (c(theta0, theta1) if there is non-differential measurement error
and c(theta00, theta01, theta10, theta11) if there is differential
measurement error) and a non-mandatory 'vcov': the variance--covariance matrix
of the coefficients}

\item{substitute}{a vector containing the error-prone measure}

\item{n}{number of independent variables in the regression model. Defaults to
one. If MeasErrorMatrix is used inside \link[mecor]{mecor}, n is neglected
and is calculated automatically.}
}
\value{
\code{MeasErrorMatrix} returns an object of \link[base]{class}
"MeasErrorMatrix".

An object of class \code{MeasErrorMatrix} is a list containing the substitute
variable and the coefficients of the measurement error model and the
variance--covariance matrix of the coefficients of the measurement error
model and the measurement error matrix. A \code{MeasErrorMatrix} is typically
used in \link[mecor]{mecor} in case of measurement error in the dependent
variable. The \code{MeasErrorMatrix} is implemented in \code{mecor} to
provide measurement error correction using external data and for conducting
sensitivity analyses. An MeasErrorMatrix object has attributes input (the
name of the substitute variable) and n (number of dependent variables) and
call (the matched call).
}
\description{
This function creates a measurement error matrix object, usually used as the
dependent variable in a regression if one wants to correct for the
measurement error in that variable, by defining the coefficients of the
measurement error matrix (dependent variable)
}
\examples{
## measurement error in dependent variable
data(ivs_o)
measerror_fit <- lm(Y_star ~ Y, data = ivs_o)
measerror_coef <- coefficients(measerror_fit)
measerror_vcov <- vcov(measerror_fit)
measerror_matrix <- with (ivs_o, MeasErrorMatrix(substitute = Y_star,
                                                 mefit = measerror_fit,
                                                 n = 2))
# identical to:
measerror_matrix <- with (ivs_o, MeasErrorMatrix(substitute = Y_star,
                                                 mefit = list(coef = measerror_coef),
                                                 n = 2))
measerror_matrix <- with (ivs_o, MeasErrorMatrix(substitute = Y_star,
                                                 mefit = list(coef = c(0, 0.5)),
                                                 n = 2))
}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
