% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mecor.R
\docType{package}
\name{mecor-package}
\alias{mecor}
\alias{mecor-package}
\title{mecor: a measurement error correction package}
\description{
mecor provides correction methods for measurement
errors in continuous endpoints.
}
\section{Package options}{

(...)
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LindaNab/mecor.git}
}

}
\author{
\strong{Maintainer}: Linda Nab \email{l.nab@lumc.nl}

}
