% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mecor.R
\name{mecor}
\alias{mecor}
\title{mecor: a measurement error correction package}
\usage{
mecor(formula, data = NULL, mevar, mefit, difvar = NULL, method = "rc",
  robust = FALSE, alpha = 0.05, B = 0)
}
\arguments{
\item{formula}{an object of class \link[stats]{formula} (or one that is
coerced to that class): a symbolic description of the naive
model to be fitted in an \link[stats]{lm} model.}

\item{data}{an optional data frame, list or environment (or
object coercible by as.data.frame to a data frame) containing
the variables in the model. If not found in \code{data}, the
variables are taken from \code{environment(formula)}, typically
the enviroment from which \code{mecor} is called.}

\item{mevar}{a non-empty character string specifying the variable
in \code{formula} with measurement error}

\item{mefit}{(a list of) object(s) of class \link[mecor]{mefit}
used to correct \code{mevar}}

\item{difvar}{needed if the structure of 'mefit' is 'differential',
a non-empty character string specifying the grouping variable in \code{formula}
that stands for the same difvar used in 'mefit'}

\item{method}{a character string indicating the method used to correct for
measurement error}

\item{alpha}{alpha level used to construct confidence intervals}

\item{B}{number of bootstrap samples}
}
\value{
\code{mecor} returns an object of \link[base]{class} "mecor"

An object of class \code{mecor} is a list containing the following components:

\item{coefficients}{a named vector containing the coefficients of the corrected model}
\item{stderr}{zero variance standard errors of coefficients of the corrected model}
\item{coef.nm}{a named vector containing the coefficients of the naive (uncorrected) model}
\item{rdf}{the residual degrees of freedom of the corrected model}
\item{call}{matched call}
\item{ci}{a named matrix containing the confidence intervals for the slope of the corrected model}
}
\description{
mecor provides correction methods for measurement
errors in continuous endpoints.
}
\examples{
X <- c(rep(0, 1000), rep(1, 1000))
Y <- X + rnorm(2000, 0, 1)
rm <- lm(Y ~ X)

##external calibration set
Xcal <- c(rep(0, 500), rep(1, 500))
Ycal <- Xcal + rnorm(1000, 0, 1)

##systematic measurement error (sme)
V_sme <- 1 + 2 * Y + rnorm(2000, 0, 3)
nm_sme <- lm(V_sme ~ X) ##compare with rm
Vcal_sme <- 1 + 2 * Ycal + rnorm(1000, 0, 3)
fit_sme <- mefit(formula = Vcal_sme ~ Ycal)
cm_sme <- mecor(formula = V_sme ~ X, mevar = "V_sme", mefit = fit_sme, method = "rc")

##differential measurement error (dme)
V_dme <- 2 + 2 * X + 3 * Y + 2 * X * Y + rnorm(2000,0,3)
nm_dme <- lm(V_dme ~ X) ##compare with rm
Vcal_dme <- 2 + 2 * Xcal + 3 * Ycal + 2 * Xcal * Ycal + rnorm(1000,0,3)
fit_dme <- mefit(formula = Vcal_dme ~ Ycal * Xcal, mestructure = "differential", difvar = Xcal)
cm_dme <- mecor(formula = V_dme ~ X, mevar = "V_dme", mefit = fit_dme, difvar = "X", method = "rc", robust = TRUE)

}
\references{

}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
