% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefit.R
\name{mefit}
\alias{mefit}
\title{Fitting Measurement Error Model}
\usage{
mefit(formula, data, robust = FALSE)
}
\arguments{
\item{formula}{formula of the measurement error model}

\item{data}{optional data frame}

\item{robust}{boolean indicating whether robust standard errors need to be calculated, the
"HC3" robust standard errors from \link[sandwich]{vcovHC} are used for the
heteroskedasticity-consistent estimation of the covariance matrix of the coefficient
estimates.}
}
\value{
\code{mefit} returns an object of \link[base]{class} "mefit".

An object of class \code{mefit} is lm object (the fit of the measurement error model) and
contains additionally, the following components:
\item{call}{the matched call}
\item{vcov}{the variance-covariance matrix of the parameters of the fitted measurement error model, in case
\code{robust} is TRUE, the heteroskedasticity-consistent covariance matrix is returned}
\item{me.model}{the assumed measurement error model}
}
\description{
This function fits a linear model to model the relationship between the test
measurement (error-prone measurement) and the reference measurement (true measurement)
in order to extract the measurement error parameters of the measurement
error model from a validation set.
}
\examples{
##generation of external validation set
Xval <- c(rep(0, 500), rep(1, 500))
Yval <- Xval + rnorm(1000, 0, 1)
Vval_cme <- Yval + rnorm(1000, 0, 3) #classical measurement error (cme)
Vval_sme <- 1 + 2 * Yval + rnorm(1000, 0, 3) #systematic measurement error (sme)
Vval_dme <- 2 + 2 * Xval + 3 * Yval + 2 * Xval * Yval + rnorm(1000, 0, 3 * (1 - Xval) + 2 * Xval) #differential measurement error (dme)

#classical
fitCme <- mefit(MeasError(Vval_cme, Yval) ~ - 1)
#systematic
fitSme <- mefit(MeasError(Vval_sme, Yval) ~ 1)
#differential
#fitDme <- mefit(MeasError(Vval_dme, Yval) ~ Xval)

}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
