% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs}
\alias{rs}
\title{Replicates Study}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
  \item{Y}{outcome, continuous}
  \item{X_star_1}{first replicate of error-prone exposure, continuous}
  \item{X_star_2}{second replicate of error-prone exposure, continuous}
  \item{X_star_3}{third replicate of error-prone exposure, continuous}
  \item{Z1}{covariate, continuous}
  \item{Z2}{covariate, continuous}
}
}
\usage{
rs
}
\description{
A simulated dataset containing 1000 observations of the outcome Y, three
replicate measures of the error-prone exposure X1_star, X2_star and X3_star
and two covariates Z1 and Z2.
}
\examples{
data("rs", package = "mecor")
}
\keyword{datasets}
