% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-mecor.R
\name{summary.mecor}
\alias{summary.mecor}
\title{Summarizing Measurement Error Correction}
\usage{
\method{summary}{mecor}(object)
}
\arguments{
\item{object}{an object of class "mecor", a result of a call to \link[mecor]{mecor}}
}
\value{
The function \code{summary.mecor} returns a list of summary statistics of the fitted
calibration model given in \code{object} using the components \code{"call"}, \code{"size"},
\code{"rdf"}, \code{"r.squared"} and \code{"sigma"} from its argument plus

\item{coefficients}{a px4 matrix with columns for the estimated coefficient, its standard error,
t-statistic and corresponding (two-sided) p-value.}
}
\description{
\code{summary} method for class "mecor"
}
\examples{
## Continuing the mecor() example:
#coef(fitSme)
#summary(cm_sme)

}
\seealso{
The calibration model fitting function \link[mecor]{mefit}, \link[base]{summary}

Function \link[stats]{coef} will extract the matrix of coefficients with standard errors,
t-statistics and p-values
}
