% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-mefit.R
\name{summary.mefit}
\alias{summary.mefit}
\title{Summarizing Measurement Error Model fits}
\usage{
\method{summary}{mefit}(object)
}
\arguments{
\item{object}{an object of class "mefit", a result of a call to \link[mecor]{mefit}}
}
\value{
The function \code{summary.mefit} returns a list of summary statistics of the fitted
measurement error models given in \code{object}
}
\description{
\code{summary} method for class "mefit"
}
\examples{
## Continuing the mefit() example:
#coef(fit1)
#summary(fit1)

}
\seealso{
The measurement error model fitting function \link[mecor]{mefit}, \link[base]{summary}
}
