% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{systematic}
\alias{systematic}
\title{Run calibration}
\usage{
systematic(formula, data = NULL, alpha = 0.05)
}
\arguments{
\item{formula}{formula the calibration model}

\item{data}{data}

\item{alpha}{alpha level}
}
\value{
This function returns the calibration parameters form the validation set.
}
\description{
This is a function that runs a calibration to substract
the measurement error parameters of the measurement error model
from the external validation set.
}
\examples{
Xcal <- c(rep(0,50),rep(1,50))
Ycal <- Xcal + rnorm(100,0,3)
Vcal <- 1 + 2*Ycal + rnorm(100,0,3)
syst <- systematic(formula = Vcal ~ Ycal + Xcal)

Vcal_dme <- 1 + (2 - 1) * Xcal + (2 - 2 * Xcal + 3 * Xcal) * Ycal + rnorm(100,0,3)
formula1 <- Vcal_dme[1:50] ~ Ycal[0:50]
formula2 <- Vcal_dme[51:100] ~ Ycal[51:100]
syst1 <- systematic(formula = formula1)
syst2 <- systematic(formula = formula2)

}
