% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uaetrial}
\alias{uaetrial}
\title{Data of a randomised controlled trial}
\format{A data frame with 100 rows and 2 variables:
\describe{
  \item{X}{Exposure}
  \item{W}{Surrogate endpoint}
}}
\usage{
uaetrial
}
\description{
A dataset containing data of a randomised controlled trial
of 100 observations in two exposure groups and measures of the
true endpoint and surrogate endpoint with systematic measurement
error.
}
\keyword{datasets}
