% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uaetrial_cal}
\alias{uaetrial_cal}
\title{Calibration set of a randomised controlled trial}
\format{A data frame with 20 rows and 3 variables:
\describe{
  \item{Xcal}{Exposure}
  \item{Ycal}{True endpoint}
  \item{Wcal}{Surrogate endpoint}
}}
\usage{
uaetrial_cal
}
\description{
A dataset containing data of a randomised controlled trial
of 20 observations in one exposure group and measures of the
true endpoint and surrogate endpoint with systematic measurement
error.
}
\keyword{datasets}
