% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasError.R
\name{MeasError}
\alias{MeasError}
\title{Create a Measurement Error Object}
\usage{
MeasError(test, reference)
}
\arguments{
\item{test}{a vector containing error-prone measurements or a matrix
with replicate error-prone measurements}

\item{reference}{a vector containing reference measurements}
}
\value{
\code{MeasError} returns an object of \link[base]{class} "MeasError".

An object of class \code{MeasError} is a data.frame and additionally has
attributes input (the name of the test and reference variables) and call (the matched call).
}
\description{
This function creates a measurement error object, usually used in a regression where one
wants to correct for the measurement error in the measured variable using a
reference measure.
}
\examples{
##measurement error in exposure
nobs <- 1e3
Z <- rnorm(nobs, 0, 1)
X <- Z + rnorm(nobs, 0, 1)
Y <- 0.5 * X + 2 * Z + rnorm(nobs, 0, 1)
W <- X + rnorm(nobs, 0, 0.5)
X <- ifelse(rbinom(nobs, 0, 0.9) == 1, NA, X)
data <- data.frame(Z, X, W, Y)
W2 <- X + rnorm(nobs, 0, 0.5)
data2 <- data.frame(Z, W, W2, Y)

MeasError(W, X)
MeasError(cbind(W, W2), NA)
}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
