% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasError.R
\name{MeasError}
\alias{MeasError}
\title{Create a Measurement Error Object}
\usage{
MeasError(substitute, reference, replicate, type)
}
\arguments{
\item{substitute}{a vector containing the error-prone measure}

\item{reference}{a vector containing the reference measure assumed without
measurement error}

\item{replicate}{a vector or matrix with replicates of the error-prone
measure with classical measurement error. This can either be
replicates obtained by using the same measurement method as the substitute
measure (replicates study) or replicates using a different measurement method
than the substitute measure (calibration study).}

\item{type}{a character string with the study type: 'ivs' for internal
validation study (default if reference not null), 'rs' for replicates study
(default if replicate not null), 'cs' for calibration study and 'evs' for
external validation study}
}
\value{
\code{MeasError} returns an object of \link[base]{class} "MeasError".

An object of class \code{MeasError} is a list containing the substitute (and
reference) variables and additionally has attributes input (the name of the
substitute and reference variables), type (the study type that provides
information to correct for the measurement error) and call (the matched
call).
}
\description{
This function creates a measurement error object, usually used as the
independent or dependent variable in a regression if one wants to correct for
the measurement error in that variable using a reference variable.
}
\examples{
## measurement error in exposure
# internal validation study
data(ivs)
me <- with (ivs, MeasError(X_star, reference = X))
# replicates study
data(rs)
me <- with (rs, MeasError(X1_star, replicate = cbind(X2_star, X3_star)))
# calibration study
data(cs)
me <- with(cs, MeasError(X_star, replicate = cbind(X1_star, X2_star), type = "cs"))
}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
