% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasError.R
\name{MeasError}
\alias{MeasError}
\title{Create a Measurement Error Object}
\usage{
MeasError(test, reference)
}
\arguments{
\item{test}{a vector containing the error-prone measurement}

\item{reference}{a vector containing the reference measurement}
}
\value{
\code{MeasError} returns an object of \link[base]{class} "MeasError".

An object of class \code{mefit} is a formula and additionally has attributes called input
(vectors containing the test and reference variable) and call (the matched call).
}
\description{
This function creates a measurement error object, usually used in a regression where one
wants to correct for the measurement error in the measured variable using a
reference measure.
}
\examples{
##generation of external validation set
Xval <- c(rep(0, 5), rep(1, 5))
Yval <- Xval + rnorm(10, 0, 1)
Vval_cme <- Yval + rnorm(10, 0, 3) #classical measurement error (cme)
Vval_sme <- 1 + 2 * Yval + rnorm(10, 0, 3) #systematic measurement error (sme)
Vval_dme <- 2 + 2 * Xval + 3 * Yval + 2 * Xval * Yval + rnorm(10, 0, 3 * (1 - Xval) + 2 * Xval) #systematic differential measurement error (dme)

cme <- MeasError(test = Vval_cme, reference = Yval)
sme <- MeasError(test = Vval_sme, reference = Yval)
dme <- MeasError(test = Vval_dme, reference = Yval)
}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
