% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mecor.R
\name{mecor}
\alias{mecor}
\title{mecor: a measurement error correction package}
\usage{
mecor(formula, data = NULL, cv, correction)
}
\arguments{
\item{formula}{formula of the naive regression model}

\item{data}{data}

\item{cv}{a non-empty character string specifying the variable
in \code{formula} to correct}

\item{correction}{the method to correct \code{cv},
a object of type \code{systematic} is implemented}
}
\value{
This function returns the corrected effect estimator.
}
\description{
mecor provides correction methods for measurement
errors in continuous endpoints.
}
\examples{
X <- c(rep(0,100),rep(1,100))
Y <- X + rnorm(200,0,3)
V <- 1 + 2*Y + rnorm(100,0,3)
model_mecor <- mecor(formula=V~X, cv = "V", correction = sys)


}
