% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefit.R
\name{mefit}
\alias{mefit}
\title{This function fits a linear regression to your validation dataset
in order to extract the measurement error parameters of the measurement
error model from the external validation set.}
\usage{
mefit(formula, data = NULL, memodel = "systematic", id = NULL)
}
\arguments{
\item{formula}{an object of class \link[stats]{formula} (or one that is
coerced to that class): a symbolic description of the naive
model to be fitted in an \link[stats]{lm} model.}

\item{data}{an optional data frame, list or environment (or
object coercible by as.data.frame to a data frame) containing
the variables in the model. If not found in \code{data}, the
variables are taken from \code{environment(formula)}, typically
the enviroment from which \code{mefit} is called.}

\item{memodel}{a character string indicating the underlying measurement model
of your data, i.e. "classical", "systematic" or "differential".}

\item{id}{indicates the grouping variable if \code{memodel} is "differential".}
}
\value{
\code{mefit} returns an object of \link[base]{class} "mefit".

An object of class \code{mefit} is a list containing the following components:

\item{coefficients}{a named vector containing the coefficients of the calibration model}
\item{model}{an \code{lm} object}
}
\description{
This function fits a linear regression to your validation dataset
in order to extract the measurement error parameters of the measurement
error model from the external validation set.
}
\examples{
Xcal <- c(rep(0,500),rep(1,500))
Ycal <- Xcal + rnorm(1000,0,3)
Vcal <- 1 + 2 * Ycal + rnorm(1000,0,3)
fit_sme <- mefit(formula = Vcal ~ Ycal, memodel = "systematic")

Vcal_dme0 <- 1 + 2 * Ycal[1:500] + rnorm(Ycal[1:500], 0, 3)
Vcal_dme1 <- 2 + 3 * Ycal[501:1000] + rnorm(Ycal[501:1000], 0, 3)
Vcal_dme <- c(Vcal_dme0, Vcal_dme1)
fit_dme <- mefit(formula = Vcal_dme ~ Ycal, memodel = "differential", id = Xcal)

}
\references{

}
\author{
Linda Nab, \email{l.nab@lumc.nl}
}
