% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{systematic}
\alias{systematic}
\title{Run calibration}
\usage{
systematic(formula, data = NULL)
}
\arguments{
\item{formula}{formula the calibration model}

\item{data}{data}

\item{type}{form of measurement error,
\code{systematic} and \code{differential} are implemented}

\item{method}{method to estimate standard errors, \code{zero-variance},
\code{delta}, \code{fieller} and \code{bootstrap} are implemented}
}
\value{
This function returns the calibration parameters.
}
\description{
This is a function that runs a calibration to substract
the measurement error parameters of the measurement error model
from the external validation set.
}
\examples{
Xcal <- c(rep(0,50),rep(1,50))
Ycal <- Xcal + rnorm(100,0,3)
Vcal <- 1 + 2*Ycal + rnorm(100,0,3)
sys <- systematic(formula = Vcal~Ycal)

Vcal_dme <- 1 + (2-1)*Xcal + (2-2*Xcal+3*Xcal)*Ycal + rnorm(100,0,3)
model_calibration1 <- lm(Vcal_dme[1:50]~Ycal[0:50])
model_calibration2 <- lm(Vcal_dme[51:100]~Y+cal[51:100])
calibration(model = list(model_calibration1,model_calibration2), form="differential")

}
